Move wrapper code to util package
