+++ FtcRobotController/src/main/java/org/firstinspires/ftc/team7316/util/AxisWrapper.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.team7316;
+package org.firstinspires.ftc.team7316.util;
 
 import android.util.Log;
