+++ FtcRobotController/src/main/java/org/firstinspires/ftc/team7316/util/GamepadButton.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.team7316;
+package org.firstinspires.ftc.team7316.util;
 
 /**
  * Created by andrew on 9/15/16.
