+++ FtcRobotController/src/main/java/org/firstinspires/ftc/team7316/util/GamepadWrapper.java
@@ -1,6 +1,4 @@
-package org.firstinspires.ftc.team7316;
-
-import android.widget.Button;
+package org.firstinspires.ftc.team7316.util;
 
 import com.qualcomm.robotcore.hardware.Gamepad;
