+++ FtcRobotController/src/main/java/org/firstinspires/ftc/team7316/util/ServoWrapper.java
@@ -1,6 +1,5 @@
-package org.firstinspires.ftc.team7316;
+package org.firstinspires.ftc.team7316.util;
 
-import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 
 /**
